requirements.txt
PySide6
psycopg2-binary
bcrypt>=4.0
Pillow
pillow-avif-plugin
